GRF-377: Separation of query parameters and query into 2 separate services
