+++ src/Akeneo/Category/back/Application/Query/GetCategoriesParametersBuilder.php
@@ -4,18 +4,16 @@ declare(strict_types=1);
 
 namespace Akeneo\Category\Application\Query;
 
-use Akeneo\Category\Domain\Model\Enrichment\Category;
-
 /**
  * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface GetCategoriesInterface
+interface GetCategoriesParametersBuilder
 {
     /**
-     * @param array<string, string> $parameters
+     * @param array<string> $categoryCodes
      *
-     * @return array<Category>
+     * @return array<string, string>
      */
-    public function execute(array $parameters): array;
+    public function build(array $categoryCodes, int $limit, int $offset, bool $isEnrichedAttributes): array;
 }
