+++ src/Akeneo/Category/back/Infrastructure/Controller/ExternalApi/ListCategoriesController.php
@@ -3,6 +3,7 @@
 namespace Akeneo\Category\Infrastructure\Controller\ExternalApi;
 
 use Akeneo\Category\Application\Query\GetCategoriesInterface;
+use Akeneo\Category\Application\Query\GetCategoriesParametersBuilder;
 use Akeneo\Category\ServiceApi\ExternalApiCategory;
 use Akeneo\Platform\Bundle\FeatureFlagBundle\FeatureFlags;
 use Akeneo\Tool\Component\Api\Exception\PaginationParametersException;
@@ -26,6 +27,7 @@ class ListCategoriesController extends AbstractController
         private readonly PaginatorInterface $paginator,
         private readonly ParameterValidatorInterface $parameterValidator,
         private readonly FeatureFlags $featureFlags,
+        private readonly GetCategoriesParametersBuilder $parametersBuilder,
         private readonly GetCategoriesInterface $getCategories,
         private readonly array $apiConfiguration
     ) {
@@ -65,16 +67,15 @@ class ListCategoriesController extends AbstractController
         if (null === $searchFilters) {
             throw new BadRequestHttpException('The search query parameter must be a valid JSON.');
         }
-        //TODO: Take limit, offset & order into account. https://akeneo.atlassian.net/browse/GRF-538
         $offset = $queryParameters['limit'] * ($queryParameters['page'] - 1);
         try {
-            //TODO: Call Filtering service (to be created) instead. https://akeneo.atlassian.net/browse/GRF-538
-            $categories = $this->getCategories->afterOffset(
-                $searchFilters,
-                $queryParameters['limit'],
-                $offset,
-                $request->query->getBoolean('with_enriched_attributes')
+            $queryParameters = $this->parametersBuilder->build(
+                    $searchFilters,
+                    $queryParameters['limit'],
+                    $offset,
+                    $request->query->getBoolean('with_enriched_attributes')
             );
+            $categories = $this->getCategories->execute($queryParameters);
         } catch (\InvalidArgumentException $exception) {
             throw new BadRequestHttpException($exception->getMessage(), $exception);
         }
