+++ src/Akeneo/Category/back/Infrastructure/Symfony/Resources/config/controllers.yml
@@ -80,6 +80,7 @@ services:
             - '@pim_api.pagination.offset_hal_paginator'
             - '@pim_api.pagination.parameter_validator'
             - '@feature_flags'
+            - '@Akeneo\Category\Application\Query\GetCategoriesParametersBuilder'
             - '@Akeneo\Category\Application\Query\GetCategoriesInterface'
             - '%pim_api.configuration%'
         calls:
