+++ src/Akeneo/Category/back/Infrastructure/Symfony/Resources/config/queries.yml
@@ -3,7 +3,6 @@ services:
     class: Akeneo\Category\Infrastructure\Storage\Sql\GetCategorySql
     arguments:
       - '@database_connection'
-      -
 
   Akeneo\Category\Application\Query\GetCategoriesInterface:
     class: Akeneo\Category\Infrastructure\Storage\Sql\GetCategoriesSql
@@ -49,3 +48,6 @@ services:
     class: Akeneo\Category\Infrastructure\Storage\Sql\GetCategoryTreeByCategoryTemplateSql
     arguments:
       - '@database_connection'
+
+  Akeneo\Category\Application\Query\GetCategoriesParametersBuilder:
+    class: Akeneo\Category\Infrastructure\Storage\Sql\GetCategoriesParametersBuilderSql
