+++ src/Akeneo/Category/back/ServiceApi/ExternalApiCategory.php
@@ -125,9 +125,4 @@ class ExternalApiCategory
             'values' => $this->getValues(),
         ];
     }
-
-    public static function fromArray(): self
-    {
-
-    }
 }
